changed error return type

+++ 0x03-asynchronous/4-get_sum.dart
@@ -14,6 +14,6 @@ Future<double> calculateTotal() async {
         
         return totalPrice;
     } catch (e) {
-        return -1;
+        return -1 as int;
     }
 }
